Renames test support files to indicate what they do
